Create Federal Solar Renewable Energy Credits Proposal.md
Solar Renewable Energy Credits (SRECs) are a financial instrument available in some jurisdictions to owners of renewable energy technologies that are interconnected to the power grid. 

The purpose of this legislation is to provide this incentive on a national level in order to spur the increased decentralization of electricity production in the United States.

The current exclusive customer of SRECs is utility companies who have mandates that a portion of their electrical generation must come from renewable sources. 

This legislation could be improved by expanding the market to corporations by requiring that a portion of their electricity consumption must come from renewable sources.